+++ Tests/SwifQLTests/SwifQLTestCase.swift
@@ -1,0 +1,21 @@
+import XCTest
+@testable import SwifQL
+
+class SwifQLTestCase: XCTestCase {
+    struct CarBrands: Codable, Tableable {
+        var id: UUID
+        var name: String
+        var createdAt: Date
+    }
+    
+    let cb = CarBrands.as("cb")
+    
+    func checkAllDialects(_ query: SwifQLable, pg: String? = nil, mySQL: String? = nil) {
+        if let pg = pg {
+            XCTAssertEqual(query.prepare(.psql).plain, pg)
+        }
+        if let mySQL = mySQL {
+            XCTAssertEqual(query.prepare(.mysql).plain, mySQL)
+        }
+    }
+}
